Extract ReferenceAdapter from ReferencesFragment
Created `ReferenceAdapter.kt` and moved the inner class out of `ReferencesFragment.kt` to clean up the fragment code. Tests pass.
